HOCS-2234: Add the SomuItemDto and core object
To allow for the use of the somu object this commit adds the DTO that will retrieve the incoming argument and parse it from the json. This specifies that we use just the uuid, somu_uuid and the data. The data is an object as it can take any form, this also allows Elastic Search to treat it as dynamic.
This commit also adds the test cases for the somu item.