HOCS-2234: update case data to contain somu item
Add a set of somu items to the case data class, this will allow us to store this under the case_data mapping on Elastic Search. This commit also adds the ability to add, remove and update a somu item.
This commit also adds the test cases for the case data object.